Move ExpiredJobReaper to reaper package
